+++ scripts/dzVents/runtime/integration-tests/stage1.lua
@@ -1007,22 +1007,14 @@ end
 local testLastUpdates = function()
 
 	local Time = require('Time')
-
 	local results = true
-
 	local now = dz.time.secondsSinceMidnight
-
---	print('Now ' .. tostring(now))
 
 	results = dz.devices().reduce(function(acc, device)
 
 		if (device.name ~= 'endResult' and device.name ~= 'stage1Trigger' and device.name ~= 'stage2Trigger') then
 
---			print('devlup ' .. tostring(device.lastUpdate.secondsSinceMidnight))
-
 			local delta = now - device.lastUpdate.secondsSinceMidnight
-
---			print('delta ' .. tostring(delta))
 
 			-- test if lastUpdate for the device is close to domoticz time
 			acc = acc and (delta <= 1)